Rename walletState -> channelState
